Building on MinGW results in build noise.

+++ src/asn1/per_support.c
@@ -6,6 +6,7 @@
 #include <asn_system.h>
 #include <asn_internal.h>
 #include <per_support.h>
+#include <stdint.h>
 
 char *
 per_data_string(asn_per_data_t *pd) {
@@ -15,7 +16,7 @@ per_data_string(asn_per_data_t *pd) {
 	snprintf(buf[n], sizeof(buf[n]),
 		"{m=%ld span %+ld[%d..%d] (%d)}",
 		(long)pd->moved,
-		(((long)pd->buffer) & 0xf),
+		(((long)(intptr_t)pd->buffer) & 0xf),
 		(int)pd->nboff, (int)pd->nbits,
 		(int)(pd->nbits - pd->nboff));
 	return buf[n];
@@ -107,7 +108,7 @@ per_get_few_bits(asn_per_data_t *pd, int nbits) {
 	ASN_DEBUG("  [PER got %2d<=%2d bits => span %d %+ld[%d..%d]:%02x (%d) => 0x%x]",
 		(int)nbits, (int)nleft,
 		(int)pd->moved,
-		(((long)pd->buffer) & 0xf),
+		(((long)(intptr_t)pd->buffer) & 0xf),
 		(int)pd->nboff, (int)pd->nbits,
 		((pd->buffer != NULL)?pd->buffer[0]:0),
 		(int)(pd->nbits - pd->nboff),
